Tag Information
Tag Information
- Added a table to show all tags and the number of videos they appear
in. The table is sortable and searchable. In reference to #9.
- Added a table to show all tags and the number of videos they appear in
for a certain month/year, sorted by most used ascending. The table is
searchable. In reference to #10.
- Added a table to show all tags and the number of videos they appear in
for a certain year, sorted by most used ascending. The table is
searchable. In reference to #11.

Miscellaneous
- Videos are now sorted in a dictionary with the key being the video's
ID. This was to make searching for a video instantenous for something I
tried but didn't need, but it may be useful in the future.