Full Dataset Overlap Histograms
Full Dataset Overlap Histograms
- Added a histogram to show the tag/title overlap percentages for each
video in a specific month in reference to #21.
- Added a histogram to show the tag/title overlap percentages for each
video in a specific year in reference to #22.
- Added a histogram to show the tag/description overlap percentages for
each video in a specific month in reference to #23.
- Added a histogram to show the tag/description overlap percentages for
each video in a specific year in reference to #24.
- Added a histogram to show the tag/title|description overlap
percentages for each video in a specific month in reference to #25.
- Added a histogram to show the tag/title|description overlap
percentages for each video in a specific year in rererence to #26.
- Added a histogram to show the percentage of words in the title used as
tags for each video in a specific month in reference to #27.
- Added a histogram to show the percentage of words in the title used as
tags for each video in a specific month in reference to #28.
- Added a histogram to show the percentage of words in the description
used as tags for each video in a specific month in reference to #29.
- Added a histogram to show the percentage of words in the description
used as tags for each video in a  specific month in reference to #30.
- Added a histogram to show the percentage of words in the union of the
title and description used as tags for each video in a specific month in
reference to #31.
- Added a histogram to show the percentage of words in the union of the
title and description used as tags for each video in a specific month in
reference to #32.

Miscellaenous Changes:
- Split up the application further by parts to make the code easier to
look at. This involved adding more options in the navigation, and more
views and controllers for specific areas.
- Added animation to views entering and leaving for a smoother user
experience.
- Added framework for the final section of this application - small
dataset overlap statistics.