Small Dataset Overlap Percentages
Small Dataset Overlap Percentages
- Added a table to show the overlap percentage between tags and
transcript in reference to #33.
- Added a table to show the overlap percentage between tags and the
union of the title, desription, and transcript in reference to #34.
- Added a table to show the percentage of words in the transcript used
as tags in reference to #35.
- Added a table to show the percentage of words in the union of the
title, description, and transcript of a video used as tags in reference
to #36.

Video Details Modal
- Clicking on a video ID in the Video Dump areas of the application will
bring up a modal showing the title, description, tags, and transcript of
the video where applicable. The video itself is also embedded in the
modal. This will help with manually confirming results, or debugging.

Miscellanous Changes:
- Added a string extension to extract text between two tags since that
didn't exist in the language already for some reason.
- Added a filter to do something that angularjs needs for a URL to
accept it for an embed..