Small Dataset Overlap Histograms
Small Dataset Overlap Histograms
- Added a histogram to show the percentage of tags in the transcript of
a video for a specific month in reference to #38.
- Added a histogram to show the percentage of tags in the transcript of
a video for a specific year in reference to #39.
- Added a histogram to show the percentage of tags used in the union of
the title, description, and transcript of a video for a specific month
in reference to #40.
- Added a histogram to show the percentage of tags used in the union of
the title, description, and transcript of a video for a specific year in
reference to #41.
- Added a histogram to show the percentage of words in the transcript of
a video found in its tags for a specific month in reference to #42.
- Added a histogram to show the percentage of words in the transcript of
a video found in its tags for a specific year in reference to #43.
- Added a histogram to show the percentage of words in the union of the
title, description, and transcript of a video found in its tags for a
specific month in reference to #44.
- Added a histogram to show the percentage of words in the union of the
title, description, and transcript of a video found in its tags for a
specific year in reference to #45.

Miscellaneous:
- Re-styled the nagivation menu to match the rest of the application.
- Added default dates to all selects to default them to the first
instance of actual data (manually).